Fix incorrect reporting of finalizable objects
The CLR 4.0+ heap details API incorrectly reported the memory range for the finalization queue vs. the f-reachable queue. The exact source of the bug was confirmed by inspecting the values returned by the DAC, and cross-checking with what SOS does. Specifically, the f-reachable queue lies in `fill_pointers[3]` to `fill_pointers[5]` (also known as `SegQueue(CriticalFinalizerListSeg)` and `SegQueueLimit(FinalizerListSeg)`), whereas the finalization queue lies in fill_pointers[0] to fill_pointers[3]. This essentially means that the `EnumerateFinalizableObjects` and `EnumerateFinalizerQueueObjectAddresses` APIs never worked correctly for CLR 4.0+.